# apps/backend/user-service/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"IUser": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"sub": {
						"type": "string"
					},
					"googleSub": {
						"type": "string"
					},
					"facebookSub": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"profile": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"PaginationResponse_IUser_": {
				"properties": {
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"totalItems",
					"totalPages",
					"currentPage"
				],
				"type": "object",
				"additionalProperties": {
					"anyOf": [
						{
							"items": {
								"$ref": "#/components/schemas/IUser"
							},
							"type": "array"
						},
						{
							"type": "number",
							"format": "double"
						}
					]
				}
			},
			"UsersPaginatedResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/PaginationResponse_IUser_"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserProfileResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/IUser"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"mongoose.Types.ObjectId": {
				"type": "string"
			},
			"UserCreationRequestParams2": {
				"properties": {
					"sub": {
						"type": "string"
					},
					"googleSub": {
						"type": "string"
					},
					"facebookSub": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"profile": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"favorites": {
						"items": {
							"$ref": "#/components/schemas/mongoose.Types.ObjectId"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"lastActive": {
						"type": "string",
						"format": "date-time"
					},
					"lastSeen": {
						"type": "string",
						"format": "date-time"
					},
					"sessions": {
						"items": {
							"properties": {
								"lastLogin": {
									"type": "string",
									"format": "date-time"
								},
								"ipAddress": {
									"type": "string"
								},
								"deviceId": {
									"type": "string"
								}
							},
							"required": [
								"lastLogin",
								"ipAddress",
								"deviceId"
							],
							"type": "object"
						},
						"type": "array"
					},
					"privacySettings": {
						"properties": {
							"profilePhotoVisibleTo": {
								"type": "string",
								"enum": [
									"everyone",
									"contacts",
									"nobody"
								]
							},
							"lastSeenVisibleTo": {
								"type": "string",
								"enum": [
									"everyone",
									"contacts",
									"nobody"
								]
							}
						},
						"required": [
							"profilePhotoVisibleTo",
							"lastSeenVisibleTo"
						],
						"type": "object"
					},
					"contacts": {
						"items": {
							"$ref": "#/components/schemas/mongoose.Types.ObjectId"
						},
						"type": "array"
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasicParams": {
				"properties": {
					"surname": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"career": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"dob": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phonenumber": {
						"type": "string"
					},
					"martial": {
						"type": "string"
					}
				},
				"required": [
					"surname",
					"lastname",
					"career",
					"email",
					"dob",
					"address",
					"phonenumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SkillParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"percent": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"percent"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExpertiseParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"proficiency": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"proficiency"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LanguageParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"proficiency": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"proficiency"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EducationParams": {
				"properties": {
					"academic": {
						"type": "string"
					},
					"school": {
						"type": "string"
					},
					"major": {
						"type": "string"
					},
					"year": {
						"type": "string"
					}
				},
				"required": [
					"academic",
					"school",
					"major",
					"year"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExperienceParams": {
				"properties": {
					"position": {
						"type": "string"
					},
					"company": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"year": {
						"type": "string"
					}
				},
				"required": [
					"position",
					"company"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReferenceParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"career": {
						"type": "string"
					},
					"company": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phonenumber": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"career",
					"company",
					"email",
					"phonenumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DescriptionParams": {
				"properties": {
					"description": {
						"type": "string"
					},
					"strength": {
						"type": "string"
					}
				},
				"required": [
					"description",
					"strength"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IUserProfile": {
				"properties": {
					"basic": {
						"$ref": "#/components/schemas/BasicParams"
					},
					"skills": {
						"items": {
							"$ref": "#/components/schemas/SkillParams"
						},
						"type": "array"
					},
					"expertise": {
						"items": {
							"$ref": "#/components/schemas/ExpertiseParams"
						},
						"type": "array"
					},
					"languages": {
						"items": {
							"$ref": "#/components/schemas/LanguageParams"
						},
						"type": "array"
					},
					"educations": {
						"items": {
							"$ref": "#/components/schemas/EducationParams"
						},
						"type": "array"
					},
					"experiences": {
						"items": {
							"$ref": "#/components/schemas/ExperienceParams"
						},
						"type": "array"
					},
					"references": {
						"items": {
							"$ref": "#/components/schemas/ReferenceParams"
						},
						"type": "array"
					},
					"descriptions": {
						"$ref": "#/components/schemas/DescriptionParams"
					},
					"certificates": {
						"items": {
							"$ref": "#/components/schemas/CertificateParams"
						},
						"type": "array"
					},
					"portfolio": {
						"items": {
							"$ref": "#/components/schemas/PortfolioParams"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"UnionProfileType": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/IdParams"
					},
					{
						"items": {
							"$ref": "#/components/schemas/IUserProfile"
						},
						"type": "array"
					},
					{
						"$ref": "#/components/schemas/BasicParams"
					},
					{
						"items": {
							"$ref": "#/components/schemas/SkillParams"
						},
						"type": "array"
					},
					{
						"items": {
							"$ref": "#/components/schemas/ExperienceParams"
						},
						"type": "array"
					},
					{
						"items": {
							"$ref": "#/components/schemas/EducationParams"
						},
						"type": "array"
					},
					{
						"items": {
							"$ref": "#/components/schemas/LanguageParams"
						},
						"type": "array"
					},
					{
						"$ref": "#/components/schemas/DescriptionParams"
					},
					{
						"items": {
							"$ref": "#/components/schemas/ReferenceParams"
						},
						"type": "array"
					},
					{
						"items": {
							"$ref": "#/components/schemas/PortfolioParams"
						},
						"type": "array"
					},
					{
						"items": {
							"$ref": "#/components/schemas/CertificateParams"
						},
						"type": "array"
					}
				],
				"nullable": true
			},
			"UserUpdateRequestParams": {
				"properties": {
					"sub": {
						"type": "string"
					},
					"googleSub": {
						"type": "string"
					},
					"facebookSub": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"profile": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CvStyleParams": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"style": {
						"type": "string"
					},
					"json": {
						"properties": {
							"objects": {
								"items": {},
								"type": "array"
							},
							"version": {
								"type": "string"
							}
						},
						"required": [
							"objects",
							"version"
						],
						"type": "object"
					}
				},
				"required": [
					"_id",
					"style",
					"json"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CorporateProfileResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"jobs": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"user": {
								"$ref": "#/components/schemas/ICorporateModel"
							}
						},
						"required": [
							"jobs",
							"user"
						],
						"type": "object"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "user-service",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/v1/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsersPaginatedResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfileResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationRequestParams2"
							}
						}
					}
				}
			}
		},
...
		"/v1/users/cv": {
			"get": {
				"operationId": "GetCvFiles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "UpdateCvFiles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"url": {
										"type": "string"
									}
								},
								"required": [
									"url"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/users/profile-detail": {
			"get": {
				"operationId": "GetProfileByID",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/IUserProfile"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUserProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/UnionProfileType"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserProfile"
							}
						}
					}
				}
			}
		},
		"/v1/users/me/photo": {
			"put": {
				"operationId": "ChangePhoto",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfileResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"photo": {
										"type": "string"
									}
								},
								"required": [
									"photo"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/users/me/favorites": {
			"post": {
				"operationId": "AddFavorite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfileResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"jobId": {
										"type": "string"
									}
								},
								"required": [
									"jobId"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetFavorites",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/v1/users/me/favorites/{jobId}": {
			"delete": {
				"operationId": "RemoveFavorite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfileResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "jobId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/users/{userId}": {
			"get": {
				"operationId": "GetUserProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfileResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfileResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateRequestParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUserById",
				"responses": {
					"204": {
						"description": "Delete Successful"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/users/getMulti/Profile": {
			"get": {
				"operationId": "GetMultiProfileUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"usersProfile": {
											"items": {
												"properties": {
													"name": {
														"type": "string"
													},
													"profile": {
														"type": "string"
													},
													"_id": {
														"type": "string"
													}
												},
												"type": "object"
											},
											"type": "array"
										}
									},
									"required": [
										"usersProfile"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "usersId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/users/uploadFile": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/v1/users/cv/{cvId}": {
			"delete": {
				"operationId": "DeleteCvFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cvId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/users/cvstyle/{style}": {
			"get": {
				"operationId": "GetCVStyle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/CvStyleParams"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "style",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/corporate/{corporateId}": {
			"get": {
				"operationId": "GetCorporateBySub",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CorporateProfileResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Corporate"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "corporateId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}